multi: pull some bin configs from integration tests into wallet
